# dmf-elearning-platform/infrastructure/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: DMF German Learning API
  version: 1.0.0
  description: |
    API for the DMF German eLearning Platform.
    Supports vocabulary, reading, listening, speaking, writing, gamification,
    education (CEFR), AI grading, and role-based admin/teacher/mentor access.
  contact:
    name: DMF Team
    url: https://dmf-elearning.com
servers:
  - url: https://api.dmf-elearning.com/api
    description: Production
  - url: https://staging-api.dmf-elearning.com/api
    description: Staging
  - url: http://localhost:3001/api
    description: Local development
...
paths:
...
  /gamification/xp:
    get:
      tags: [Gamification]
      summary: Get user XP and level
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: XP data
  /gamification/xp/award:
    post:
      tags: [Gamification]
      summary: Award XP to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, source]
              properties:
                userId: { type: string }
                amount: { type: integer }
                source: { type: string }
      responses:
        '200':
          description: XP awarded
...
components:
  securitySchemes:
    rbac:
      type: apiKey
      in: query
      name: userId
      description: User ID for role-based access control